qoh tests
